+++ CHANGELOG.md
@@ -6,6 +6,12 @@ Changes:
 
 \-
 
+## v2023-05-04-1
+
+Changes:
+
+- `flux-button-group`
+
 ## v2023-04-11-1
 
 Changes: